the correction moved in the witness and stayed, with an Assert around it, in the witness's test
The fourth round, on the repair rather than on the change, found the shape this
repo names most often. SeatCutBlindnessWitness had its soundness argument
corrected: the bad-reduction coincidence needs no coupling divisible by both
ranking primes, only different bonds divisible by different ranking primes, each
near 10^9 and accepted. Three surfaces carried the old argument and one moved.
The one left behind was SeatCutBlindnessWitnessTests, where it was not merely
stale but GATED, by an assert comparing two compile-time constants that cannot
fail, inside the green suite the previous commit cited as its evidence.

That assert is now the measurement. N = 3, XY, bonds equal to the witness's own
two ranking primes read one blind direction where SeatCutBlindnessClaim.Blind
says zero, and zero again one step off either bond, so the Diagnostics half
certifies its own correction from below instead of inheriting MirrorWorld's.

Three more from the same round. The corrected sentence over-claimed in the other
direction, a one-sidedness holding "at any coupling", which walks straight back
into the 2026-08-30 caught error where the count was measured coming out too
SMALL past the magnitude guard; the premise is named there now. Calling
MirrorWorld and Diagnostics "the two halves of the typed layer" was wrong in
three places at once: MirrorWorld references no RCPsiSquared.* project by design
and is not in that layer, whose halves are the Claim graph in Core and the
witnesses in Diagnostics. And F157 told the disjointness story stronger than the
repo's reduction graph; it now says that the live C# witnesses are a third route
and share 2147483647 with the Python one.

One reviewer finding did not survive being acted on, and the ledger keeps it
because the repair is where the risk is. PrimeFactors computing q*q in int was
reported as this repo's own int-overflow trap living inside the primitive built
to guard that trap. The product does wrap, but m % q stays an honest
divisibility test at every q, so no false divisor is ever reported and the answer
does not change; the loop merely scans past the root. The bound is widened to
long anyway and the reason is stated rather than gated, because the first attempt
to gate it was a test that survived its own mutation, written minutes after
repairing gates that could not fail.

552 MirrorWorld, 98 Diagnostics seat, 14 Core registry tests green.

+++ compute/MirrorWorld/ModP.cs
@@ -115,7 +115,12 @@ public static class ModP
     public static int[] PrimeFactors(int m)
     {
         var qs = new List<int>();
-        for (int q = 2; q * q <= m; q++)
+        // long, because q*q in int wraps from q = 46341 and floor(sqrt(int.MaxValue)) is 46340, so for
+        // m near the top of int an int bound goes NEGATIVE, stays <= m, and the loop scans far past the
+        // root. It is a cost and not a wrong answer: m % q is an honest divisibility test at every q,
+        // wrapped bound or not, so no false divisor is ever reported. Stated rather than gated for that
+        // reason, since the only thing a gate here could read is the clock.
+        for (int q = 2; (long)q * q <= m; q++)
             if (m % q == 0) { qs.Add(q); while (m % q == 0) m /= q; }
         if (m > 1) qs.Add(m);
         return qs.ToArray();
